use Stg_Base_SchoolTeachers to create Fact_TeacherCohort
the previous use of Fact_TeacherMobility to create Fact_TeacherCohort
was out of laziness than an actual dependence on anything in that table.
This makes use of a table further upstream that is more directly related
to the concept of teacher cohorts.

it also adds the most recent year of data to this table, which had been
missing before. that didn't matter since we can't use the most recent
year as the start year for mobility analysis